requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
locust==1.4.1
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.3
pyotp==2.3.0
python-gnupg==0.4.6
pyzmq==20.0.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
